Extract match logic to match engine, simplified match controller
