+++ app/com/codelab27/cards9/game/engines/package.scala
@@ -1,0 +1,11 @@
+package com.codelab27.cards9.game
+
+package object engines {
+
+  object fight extends FightEngine
+
+  object board extends BoardEngine
+
+  object matches extends MatchEngine
+
+}
